requirements.txt
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
googleScrapy==0.0.1
idna==2.10
Jinja2==3.0.0
joblib==1.0.1
MarkupSafe==2.0.0
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pathy==0.5.2
preshed==3.0.5
pydantic==1.7.4
pymongo==3.11.4
pyparsing==2.4.7
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
thinc==8.0.3
tqdm==4.60.0
typer==0.3.2
urllib3==1.26.4
wasabi==0.8.2
